wip: extract listener utils
